finalterm/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
msgpack==1.1.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.6.0
zope.interface==7.2
# Environment variables
python-dotenv==1.0.0
# PostgreSQL support
psycopg2-binary==2.9.9
dj-database-url==2.2.0
# Rate limiting
django-ratelimit==4.1.0
djangorestframework-stubs>=3.16.0
django-stubs[compatible]>=5.2.0
mypy>=1.0.0